Add project description (#4)
# Pull Request

## Description

This pull request includes minor updates to documentation and
configuration files. The changes enhance clarity and ensure
compatibility with the latest tools.

Documentation update:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R3-R4):
Added a description of the repository's purpose, highlighting that it
contains reusable GitHub Actions workflows for tasks like CI/CD,
testing, and deployment.

Configuration update:

*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the `min_version` of Lefthook from `1.11.11` to `1.11.12` to
ensure compatibility with the latest version.