Ignore Pinact Warnings (#31)
# Pull Request

## Description

This pull request introduces configuration and integration for Pinact, a
tool for managing reusable workflows. The changes include adding a new
configuration file and updating commands in the `Justfile` to reference
the configuration.

### Pinact Configuration:

* Added a new Pinact configuration file
`.github/other-configurations/pinact.yml` with schema validation,
versioning, and a list of reusable workflows to ignore.

### Pinact Integration:

* Updated the `Justfile` to include a `pinact_config` variable pointing
to the new configuration file and modified all Pinact-related commands
(`pinact-run`, `pinact-check`, `pinact-update`) to use this
configuration.