Add CodeQL Analysis Reusable Workflow (#35)
# Pull Request

## Description

This pull request introduces a CodeQL analysis workflow to enhance the
security and quality of the codebase. The changes include adding a new
reusable workflow for CodeQL analysis, integrating it into existing
workflows, and configuring necessary permissions and inputs.

### CodeQL Analysis Workflow Integration:

*
[`.github/workflows/codeql-analysis.yml`](diffhunk://#diff-63bd641104d10e25f141d518a16b22a151d125e12701df2f9e79734b23b90188R1-R37):
Added a new reusable workflow for CodeQL analysis, defining inputs,
secrets, permissions, and steps for initializing and performing the
analysis.

### Updates to Existing Workflows:

*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34R9):
Included the new `codeql-analysis.yml` workflow in the monitored paths
to ensure it triggers appropriately during releases.
*
[`.github/workflows/run-reusable-worklows.yml`](diffhunk://#diff-250045fa99d200d041d50f003ae73dea85ad0028df828cca968ec02bfd694ad9R37-R47):
Integrated the CodeQL analysis workflow as a job named `codeql-checks`,
specifying permissions, inputs, and secrets.