Use Reusable CodeQL Analysis in Code Checks (#38)
# Pull Request

## Description

This pull request includes updates to the GitHub Actions workflows to
improve flexibility and consistency, particularly for CodeQL analysis.
The changes include introducing a matrix strategy for language analysis,
renaming job identifiers for clarity, and standardizing terminology.

### Improvements to CodeQL workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L45-R53):
Renamed job `run-codeql-analysis` to `codeql-checks` and introduced a
matrix strategy for analyzing multiple languages. Updated the `language`
parameter to use `${{ matrix.language }}`.
*
[`.github/workflows/codeql-analysis.yml`](diffhunk://#diff-63bd641104d10e25f141d518a16b22a151d125e12701df2f9e79734b23b90188L8-R16):
Renamed job `run-codeql-analysis` to `codeql-analysis` and updated the
job name to "Analyse code" for consistency. Adjusted the description of
the `language` input to use British English spelling ("analyse").

### Updates to reusable workflows:

*
[`.github/workflows/run-reusable-workflows.yml`](diffhunk://#diff-250045fa99d200d041d50f003ae73dea85ad0028df828cca968ec02bfd694ad9R43-R46):
Added a matrix strategy to support multiple languages in the reusable
workflow for CodeQL analysis. Updated the `language` parameter to use
`${{ matrix.language }}` for dynamic input.