Bump GitHub Actions (#40)
# Pull Request

## Description

This pull request updates dependencies in GitHub Actions workflows to
ensure compatibility and access to the latest features and fixes. The
changes involve bumping versions for the CodeQL and `setup-uv` actions.

### Dependency Updates:

*
[`.github/workflows/codeql-analysis.yml`](diffhunk://#diff-63bd641104d10e25f141d518a16b22a151d125e12701df2f9e79734b23b90188L27-R33):
Updated the `github/codeql-action/init` and
`github/codeql-action/analyze` actions from version `v3.28.17` to
`v3.28.18` to include the latest improvements and fixes.

*
[`.github/workflows/common-code-checks.yml`](diffhunk://#diff-d4fd1616efd0d349042c6f234e04f6a67b1cf5c85a83ef463226a145f0c1cec6L64-R64):
Updated the `astral-sh/setup-uv` action from version `v6.0.1` to
`v6.1.0` to ensure the latest version of `uv` is installed.
[[1]](diffhunk://#diff-d4fd1616efd0d349042c6f234e04f6a67b1cf5c85a83ef463226a145f0c1cec6L64-R64)
[[2]](diffhunk://#diff-d4fd1616efd0d349042c6f234e04f6a67b1cf5c85a83ef463226a145f0c1cec6L87-R87)