Validate ${{ github.repository }} (#44)
# Pull Request

## Description

This pull request updates the cache deletion logic in the
`.github/workflows/common-clean-caches.yml` file to improve validation
and error handling for repository names.

Enhancements to cache deletion logic:

*
[`.github/workflows/common-clean-caches.yml`](diffhunk://#diff-13fad14fa3972181d9dae18fc35384d3396e52fd6e985447df6b9257982ce9cbL25-R35):
Added a validation check for the `GITHUB_REPOSITORY` variable to ensure
it matches the expected repository name format before attempting to
delete caches. If the format is invalid, an error message is logged, and
the workflow exits with a failure status.