Handle Pinact Config if exists  (#71)
# Pull Request

## Description

This pull request updates the `.github/workflows/common-code-checks.yml`
file to enhance the validation of pinned GitHub Actions. The changes
introduce conditional checks to ensure actions are pinned, both when a
configuration file is present and when it is absent.

Enhancements to GitHub Actions validation:

*
[`.github/workflows/common-code-checks.yml`](diffhunk://#diff-d4fd1616efd0d349042c6f234e04f6a67b1cf5c85a83ef463226a145f0c1cec6L94-R105):
Added two separate steps to check if actions are pinned. One step runs
when the configuration file `.github/other-configurations/pinact.yml` is
present, and another runs when the file is absent. This ensures
comprehensive validation regardless of configuration availability.