Bump CodeQL Analysis (#74)
# Pull Request

## Description

This pull request updates the CodeQL GitHub Actions version in the
`.github/workflows/codeql-analysis.yml` file. The changes ensure the
workflow uses the latest version of CodeQL for initialization and
analysis.

*
[`.github/workflows/codeql-analysis.yml`](diffhunk://#diff-63bd641104d10e25f141d518a16b22a151d125e12701df2f9e79734b23b90188L27-R33):
Updated the `github/codeql-action/init` and
`github/codeql-action/analyze` actions from version `v3.29.1` to
`v3.29.2` to leverage improvements and fixes in the newer version.