Add Gitleaks Job (#78)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow to
enhance security by checking for secrets in the codebase using Gitleaks.

Security enhancement:

*
[`.github/workflows/common-code-checks.yml`](diffhunk://#diff-d4fd1616efd0d349042c6f234e04f6a67b1cf5c85a83ef463226a145f0c1cec6R137-R154):
Introduced a new job `run-gitleaks` that uses the Gitleaks action to
scan for secrets in the repository. The job includes steps to check out
the repository and execute the Gitleaks scan with the necessary
permissions and environment variables.