Fix Markdown Formatting Error (#116)
# Pull Request

## Description

This pull request updates the documentation in `README.md` to include an
example configuration for the `common-code-checks.yml` workflow. The
main change is the addition of a code block demonstrating how to set up
the CodeQL analysis job for advanced security analysis.

Documentation update:

* Added a YAML code block to `README.md` showing an example
configuration for the `codeql-analysis` job, helping users understand
how to integrate CodeQL analysis into their workflows.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R83)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R97)