Update Setup UV Version (#160)
# Pull Request

## Description

This pull request updates the GitHub Actions workflow to use the latest
version of the `uv` setup action. This ensures that the CI pipeline
benefits from recent improvements and bug fixes in the tool.

* Updated the `astral-sh/setup-uv` action from version `v6.6.1` to
`v6.7.0` in `.github/workflows/common-code-checks.yml` to ensure the
latest features and fixes are used in CI jobs.