Update CodeQL Action (#171)
# Pull Request

## Description

This pull request updates the GitHub Actions workflows to use the latest
version of the CodeQL action (`v3.30.5`) instead of the previous version
(`v3.30.4`). This ensures that the workflows benefit from the latest
security patches and improvements provided by the CodeQL team.

**Workflow dependency updates:**

* Updated `github/codeql-action/init` and `github/codeql-action/analyze`
actions to version `v3.30.5` in `.github/workflows/codeql-analysis.yml`
for improved security and reliability.
* Updated `github/codeql-action/upload-sarif` action to version
`v3.30.5` in `.github/workflows/common-code-checks.yml` to ensure
compatibility and benefit from the latest fixes.