Update CodeQL Actions (#207)
# Pull Request

## Description

This pull request updates the version of the `github/codeql-action` used
in several GitHub Actions workflow files to v4.31.2. The main goal is to
ensure the workflows use the latest patch release for improved
stability, security, and compatibility.

**Workflow dependency updates:**

*
[`.github/workflows/codeql-analysis.yml`](diffhunk://#diff-63bd641104d10e25f141d518a16b22a151d125e12701df2f9e79734b23b90188L27-R33):
Updated both the `init` and `analyze` steps to use
`github/codeql-action` v4.31.2 instead of v4.31.0.
*
[`.github/workflows/common-code-checks.yml`](diffhunk://#diff-d4fd1616efd0d349042c6f234e04f6a67b1cf5c85a83ef463226a145f0c1cec6L138-R138):
Updated all instances of `github/codeql-action/upload-sarif` to v4.31.2
for Anchore and Trivy SARIF report uploads.
[[1]](diffhunk://#diff-d4fd1616efd0d349042c6f234e04f6a67b1cf5c85a83ef463226a145f0c1cec6L138-R138)
[[2]](diffhunk://#diff-d4fd1616efd0d349042c6f234e04f6a67b1cf5c85a83ef463226a145f0c1cec6L165-R165)
[[3]](diffhunk://#diff-d4fd1616efd0d349042c6f234e04f6a67b1cf5c85a83ef463226a145f0c1cec6L274-R274)