Update Pinact Action (#234)
# Pull Request

## Description

This pull request updates the GitHub Actions workflow to use a newer
version of the `pinact-action` for checking that actions are pinned.
This ensures the workflow benefits from the latest features and fixes in
the action.

Workflow dependency update:

* Updated `suzuki-shunsuke/pinact-action` from version `v1.0.0` to
`v1.1.0` in `.github/workflows/common-code-checks.yml` for both
configuration and non-configuration checks.